Nema 17 compatible extruder
We noticed that our 1/50 motor is too slow when extruding with
resolution of 1mm and 2mm. For that we recomend you to use a faster
motor as the nema 17. We design it so you just need to change the motors
and not the full extruder. Use an electrical wire connector near the
extruder to quick change it. All process can be done under 5min,