Extruder upgrade for Filaflex
In our original extruder design we left some open spaces for you to see
if and were the cloghs could happen.
Unfortunelly that gaps are a problem when you use filaflex.
We intruduced in ths version a PTFE guide that you can adjust to have 0
gaps but still have space for you to track any problem.